refactor: extract internal/peripheral and internal/measurement packages
Moves peripheral types, store, and service into internal/peripheral.
Introduces internal/measurement with Point/Query/Reading/Writer/Querier
so peripheral and sensors can share measurement types without circular
imports. InfluxDB client updated to implement measurement.Client.
Sensors package keeps thin aliases/wrappers for backward compat.

Closes #98
